Split restricted editing plugins.
